+++ packages/core/src/index.ts
@@ -22,7 +22,6 @@ export * from './migration';
 export * from './packs';
 export * from './validators';
 export * from './release';
-export * from './providers';
 export {
   EpicSchema as UnifiedEpicSchema,
   EpicProfileSchema,