+++ packages/extension/src/webview/v3/shell/sidebar/index.ts
@@ -1,4 +1,0 @@
-export * from './QuotaTracker';
-export * from './RecentEpics';
-export * from './TemplateChips';
-export * from './McpList';